MYR-1172: moved styles to its own file
